refactor(api): update collections pattern
