+++ src/common/api/index.js
@@ -1,1 +1,4 @@
 export { getUnleashAssignments } from './queries/get-unleash-assignments'
+
+export { getCollectionBySlug } from './queries/get-collection-by-slug'
+export { getCollections } from './queries/get-collections'
