+++ src/containers/collections/collections.state.js
@@ -1,7 +1,7 @@
 import { select, takeLatest, put } from 'redux-saga/effects'
 import { deriveCollection } from 'common/api/derivers/item'
-import { getCollectionBySlug } from 'common/api/collections'
-import { getCollections } from 'common/api/collections'
+import { getCollectionBySlug } from 'common/api'
+import { getCollections } from 'common/api'
 import { saveItem } from 'common/api/_legacy/saveItem'
 import { removeItemByUrl } from 'common/api/_legacy/removeItem'
 import { saveItems } from 'common/api/_legacy/saveItem'
